Security on user access

+++ cms/cmk-cms/view/users.php
@@ -2,39 +2,40 @@
 if ( !isset($view_files) )
 {
 	require '../config.php';
+	$view_file = 'users';
+
 
     //security('../');
-
 }
 //if session users is not defined, define it with an empty array
-if (!isset($_SESSION['users']) )    $_SESSION['users']                  = [];
+if (!isset($_SESSION[$view_file]) )    $_SESSION[$view_file]               = [];
 //if these URL params are set, save their value to session
-if (isset($_GET['page-no']))        $_SESSION['users']['page_no']       = $_GET['page-no'];
-if (isset($_GET['sort-by']))        $_SESSION['users']['sort_by']       = $_GET['sort-by'];
-if (isset($_GET['order']))          $_SESSION['users']['order']         = $_GET['order'];
+if (isset($_GET['page-no']))        $_SESSION[$view_file]['page_no']       = $_GET['page-no'];
+if (isset($_GET['sort-by']))        $_SESSION[$view_file]['sort_by']       = $_GET['sort-by'];
+if (isset($_GET['order']))          $_SESSION[$view_file]['order']         = $_GET['order'];
 
 if (isset($_GET['page-length']) && $_GET['page-length'] >= min($page_lengths) && $_GET['page-length'] <= max($page_lengths))
 {
-    $_SESSION['users']['page_length']   = $_GET['page-length'];
-    unset($_SESSION['users']['page_no']);
+    $_SESSION[$view_file]['page_length']   = $_GET['page-length'];
+    unset($_SESSION[$view_file]['page_no']);
 }
 
 // If search is defined in URL params and the value is not empty, save the value to session
 if(isset($_GET['search']) && !empty($_GET['search']))
 {
-    $_SESSION['users']['search'] = $_GET['search'];
-    unset($_SESSION['users']['page_no']);
+    $_SESSION[$view_file]['search'] = $_GET['search'];
+    unset($_SESSION[$view_file]['page_no']);
 }
 
 // If search is defined in URL params and the value is empty, unset the session to clear search
-if(isset($_GET['search']) && empty($_GET['search'])) unset($_SESSION['users']['search']);
+if(isset($_GET['search']) && empty($_GET['search'])) unset($_SESSION[$view_file]['search']);
 
 // Defaults
-$page_length    = isset($_SESSION['users']['page_length'])  ? intval($_SESSION['users']['page_length'])            : PAGE_LENGTH;
+$page_length    = isset($_SESSION[$view_file]['page_length'])  ? intval($_SESSION[$view_file]['page_length'])            : PAGE_LENGTH;
 $page_no        = isset($_SESSION['page_no'])               ? $_GET['page-no']                                     : 1;
 $sort_by	    = isset($_SESSION['sort_by'])               ? $_GET['sort-by']	                                   : 'created';
 $order			= isset($_SESSION['order'])	                ? $mysqli->escape_string($_GET['order'])	           : 'desc';
-$search         = isset($_SESSION['users']['search'])       ? $mysqli->escape_string($_SESSION['users']['search']) : '';
+$search         = isset($_SESSION[$view_file]['search'])       ? $mysqli->escape_string($_SESSION[$view_file]['search']) : '';
 $icon_created	= $icon_name = $icon_email = $icon_role = $icon_status = '';
 
 //if else to get the items to sort accordingly
@@ -73,18 +74,19 @@ switch($sort_by)
         $order_by       ="user_status " . strtoupper($order);
 		break;
 }
-// If delete and id is defined in URL params and the id is not empty, delete the selected user
-if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']))
+// If delete and id is defined in URL params, the id is not empty and id doesn't match the current users id, delete the selected user
+if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']) && ( $_GET['id'] != $_SESSION['user']['id']))
 {
     $id = intval($_GET['id']);
 
-    $id = intval($_GET['id']);
     // Get the users from the Database
     $query = "
 				SELECT 
-					user_name , user_email, fk_role_id 
+					user_name , role_access_level
 				FROM  
 					users 
+			    INNER JOIN 
+			    	roles ON users.fk_role_id = roles.role_id
 				WHERE 
 					user_id = $id";
     $result = $mysqli->query($query);
@@ -95,23 +97,31 @@ if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']))
         query_error($query, __LINE__, __FILE__);
     }
 
-    //Return the information from the Database as an object
-    $row = $result->fetch_object();
-    $query = "
+    if($result ->num_rows == 1)
+	{
+		//Return the information from the Database as an object
+		$row = $result->fetch_object();
+
+		//Only delete the selected user if the access level is below the current users access level or is Super Administrator
+		if($row->role_access_level < $_SESSION['user']['access_level']|| $_SESSION['user']['access_level'] == 1000)
+		{
+			$query = "
             DELETE FROM 
                     users
             WHERE 
                     user_id = $id";
-    $result = $mysqli->query($query);
+			$result = $mysqli->query($query);
 
-    // If result returns false, use the function query_error to show debugging info
-    if(!$result)
-    {
-        query_error($query, __LINE__, __FILE__);
-    }
+			// If result returns false, use the function query_error to show debugging info
+			if(!$result)
+			{
+				query_error($query, __LINE__, __FILE__);
+			}
+		}
 
-    //Use a function to insert event in log
-    create_event('delete', 'af brugeren' . $row->user_name, 100);
+		//Use a function to insert event in log
+		create_event('delete', 'af brugeren' . $row->user_name, 100);
+	}
 }
 ?>
 
@@ -120,7 +130,7 @@ if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']))
 	<span class="title">
 		<?php
 		// Get icon and title from Array $files, defined in config.php
-		echo $view_files['users']['icon'] . ' ' . $view_files['users']['title']
+		echo $view_files[$view_file]['icon'] . ' ' . $view_files[$view_file]['title']
 		?>
 	</span>
 </div>
@@ -173,19 +183,19 @@ if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']))
 				<thead>
 				<tr>
 					<th>
-						<a href="index.php?page=users&sort-by=created&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=created&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_created . CREATED ?></a>
+						<a href="index.php?page=<?php echo $view_file ?>&sort-by=created&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=created&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_created . CREATED ?></a>
 					</th>
 					<th>
-						<a href="index.php?page=users&sort-by=name&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=name&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_name . NAME ?></a>
+						<a href="index.php?page=<?php echo $view_file ?>&sort-by=name&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=name&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_name . NAME ?></a>
 					</th>
 					<th>
-						<a href="index.php?page=users&sort-by=email&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=email&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_email . EMAIL ?></a>
+						<a href="index.php?page=<?php echo $view_file ?>&sort-by=email&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=email&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_email . EMAIL ?></a>
 					</th>
 					<th>
-						<a href="index.php?page=users&sort-by=role&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=role&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_role . ROLE ?></a>
+						<a href="index.php?page=<?php echo $view_file ?>&sort-by=role&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=role&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_role . ROLE ?></a>
 					</th>
 					<th class="toggle">
-						<a href="index.php?page=users&sort-by=status&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=status&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_status . STATUS ?></a>
+						<a href="index.php?page=<?php echo $view_file ?>&sort-by=status&order=<?php echo $new_order ?>" data-page="users" data-params="sort-by=status&order=<?php echo $new_order ?>" title="<?php echo SORT_BY_THIS_COLUMN ?>"><?php echo $icon_status . STATUS ?></a>
 					</th>
 					<th class="icon"></th>
 					<th class="icon"></th>
@@ -204,13 +214,27 @@ if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']))
 						            LIKE '%$search%'";
                 }
 
+                $access_level_sql = '';
+
+				// If current users access level is below 1000 (not Super Administrator), add filter to sql, so users with higher access level is not visible
+                if($_SESSION['user']['access_level'] < 1000 )
+				{
+					$user_access_level = intval($_SESSION['user']['access_level']);
+
+					$access_level_sql = "
+									AND
+										role_access_level <= $user_access_level";
+				}
+
 					$query = "
 							SELECT 
-								user_id, user_status, DATE_FORMAT (user_created, '%a, %e. %b %Y kl. %H:%I') AS user_created_formatted, user_name, user_email, role_name
+								user_id, user_status, DATE_FORMAT (user_created, '" . DATETIME_FORMAT . "') AS user_created_formatted, user_name, user_email, role_name, role_access_level
 							FROM 
 								users 
 							INNER JOIN 
-								roles ON users.fk_role_id = roles.role_id $search_sql
+								roles ON users.fk_role_id = roles.role_id 
+							WHERE
+								1=1 $search_sql $access_level_sql
 							";
 					$result = $mysqli->query($query);
 
@@ -248,17 +272,39 @@ if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']))
 
                             <!-- TOGGLE TIL AKTIVER/DEAKTIVER ELEMENT -->
                             <td class="toggle">
+								<?php
+								if($row->user_id != $_SESSION['user']['id'] && ($row->role_access_level < $_SESSION['user']['access_level'] || $_SESSION['user']['id'] == 1000))
+								{ // Don't show toggle for current user or users with the same access level, unless you're Super Administrator
+								?>
                                 <input type="checkbox" class="toggle-checkbox" id="<?php echo $row->user_id ?>" data-type="users" <?php if($row->user_status == 1) { echo 'checked';} ?> >
+								<?php
+								}
+								?>
                             </td>
 
                             <!-- REDIGER LINK -->
                             <td class="icon">
-                                <a class="<?php echo $buttons['edit'] ?>" href="index.php?page=user-edit&id=<?php echo $row->user_id ?>" data-page="user-edit" data-params="id=<?php echo $row->user_id ?>" title="<?php echo EDIT_ITEM ?>"><?php echo $icons['edit'] ?></a>
+								<?php
+								//don't show edit link for users with same access level, only our own user
+								if($row->user_id == $_SESSION['user']['id'] || $row->role_access_level < $_SESSION['user']['access_level'] || $_SESSION['user']['id'] == 1000)
+								{
+								?>
+								<a class="<?php echo $buttons['edit'] ?>" href = "index.php?page=user-edit&id=<?php echo $row->user_id ?>" data - page = "user-edit" data - params = "id=<?php echo $row->user_id ?>" title = "<?php echo EDIT_ITEM ?>" ><?php echo $icons['edit'] ?></a>
+								<?php
+								}
+								?>
                             </td>
 
                             <!-- SLET LINK -->
                             <td class="icon">
-                                <a class="<?php echo $buttons['delete'] ?>" data-toggle="confirmation" href="index.php?page=users&id=<?php echo $row->user_id ?>&delete" data-page="users" data-params="id=<?php echo $row->user_id ?>&delete" title="<?php echo DELETE_ITEM ?>"><?php echo $icons['delete'] ?></a>
+								<?php
+								if($row->user_id != $_SESSION['user']['id'] && ($row->role_access_level < $_SESSION['user']['access_level'] || $_SESSION['user']['id'] == 1000))
+								{ // Don't show delete for current user or users with the same access level, unless you're Super Administrator
+								?>
+                                <a class="<?php echo $buttons['delete'] ?>" data-toggle="confirmation" href="index.php?page=<?php echo $view_file ?>&id=<?php echo $row->user_id ?>&delete" data-page="<?php echo $view_file ?>" data-params="id=<?php echo $row->user_id ?>&delete" title="<?php echo DELETE_ITEM ?>"><?php echo $icons['delete'] ?></a>
+									<?php
+								}
+									?>
                             </td>
                         </tr>
                         <?php
@@ -277,7 +323,7 @@ if (isset($_GET['delete'], $_GET['id']) && !empty($_GET['id']))
 				<ul class="pagination">
                     <?php
 
-                            pagination($page_no, $items_total, $page_length);
+                            pagination($view_file, $page_no, $items_total, $page_length);
